// src/apps/frontend/src/app/api/model/user-post-request-dto.ts
/**
 * ToDo Application API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserPostRequestDto { 
    /**
     * ユーザー名
     */
    name?: string;
    /**
     * ユーザーのメールアドレス
     */
    email: string;
    /**
     * ユーザーのパスワード
     */
    password: string;
}